backend/requirements.txt
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.29.0,<1.0.0
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0,<3.0.0
asyncpg>=0.29.0
aiosqlite>=0.19.0
alembic>=1.12.0,<2.0.0
pydantic>=2.6.0,<3.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
bcrypt>=4.0.1,<5.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
pytest-asyncio>=0.23.0
pytest>=7.4.0
httpx>=0.27.0

